requirements.txt
fastapi
uvicorn
langchain_openai
browser_use
python-dotenv
playwright